changes and travis check

+++ infrastructure/aws/cloudformation/csye6225-aws-cf-terminate-application-stack.sh
@@ -1,9 +1,20 @@
 #!/bin/bash
 StackName=$1
+DomainName=$2
 if [ -z "$StackName" ]; then
   echo "ERROR: Stackname expected....."
   exit 1
 fi
+if [ -z "$DomainName" ]; then
+  echo "No domain name provided. Script exiting.."
+  exit 1
+fi
+
+DomainName=code-deploy.$DomainName.me
+
+
+echo "Cleaning s3 bucket $DomainName"
+aws s3 rm s3://$DomainName --recursive
 
 echo "Terminating $StackName network setup"
 aws cloudformation delete-stack --stack-name $StackName
